whisper_python/requirements.txt
Flask
openai-whisper
ffmpeg-python
snowflake-ml-python